requirements.txt
google-generativeai>=0.8.4
python-dotenv==1.0.1
rich==13.9.4
typer==0.15.2
pytest>=7.0.0
black>=22.0.0
langchain-community>=0.0.13
langchain-google-genai>=2.0.11 